// src/enums/dashboard.enum.ts
export enum EnumDashboardPeriodType {
  DAY = "day",
  WEEK = "week",
  MONTH = "month",
  YEAR = "year",
  CUSTOM = "custom",
}

export enum EnumDashboardWidgetCategory {
  MONEY = "money",
  SERVICES = "services",
}

export enum EnumDashboardWidgetId {
  DAILY_PAY = "daily-pay",
  REVENUE = "revenue",
  TIPS = "tips",
  PAYMENT_METHODS = "payment-methods",
  TOP_SERVICES = "top-services",
}

export enum EnumDashboardScreen {
  MD = "md",
  SM = "sm",
}